requirements.txt
flask
pandas
numpy
python-dotenv
langchain
langchain-community
faiss-cpu
sentence-transformers
